Separate functions  into the modules
