+++ src/style.css
@@ -94,10 +94,16 @@ ul {
   color: gray;
 }
 
-.to-do-item p {
+.to-do-item .editable-input {
   flex-grow: 1;
   padding-left: 15px;
   font-weight: 300;
+  border: none;
+  font-size: 15px;
+}
+
+.editable-input:focus{
+  border: none;
 }
 
 .to-do-item span {
